+++ config/app.php
@@ -1,10 +1,11 @@
-<?php return array (
-  'name' => 'fic',
-  'version' => '1.0.1',
-  'env' => 'development',
-  'providers' => 
-  array (
-    0 => 'App\\Providers\\AppServiceProvider',
-    1 => 'Spatie\\OpenApiCli\\OpenApiCliServiceProvider',
-  ),
-);
+<?php
+
+return [
+    'name' => 'fic',
+    'version' => '1.0.1',
+    'env' => 'development',
+    'providers' => [
+        0 => 'App\\Providers\\AppServiceProvider',
+        1 => 'Spatie\\OpenApiCli\\OpenApiCliServiceProvider',
+    ],
+];